anv: Get ready for > 4GB memory objects
We were using uint32_t offsets a number of different places.  These
should all be 64 bits if we're going to allow memory objects > 2GB.
We also use negative offsets for the surface state pool so we make all
of the deltas int64_t.  This should be sufficient for anything which
fits in a 48-bit address space.  We also assert that the deltas are
in-range for what the kernel supports.

+++ src/intel/vulkan/genX_cmd_buffer.c
@@ -155,7 +155,7 @@ genX(cmd_buffer_emit_state_base_address)(struct anv_cmd_buffer *cmd_buffer)
 static void
 add_surface_state_reloc(struct anv_cmd_buffer *cmd_buffer,
                         struct anv_state state,
-                        struct anv_bo *bo, uint32_t offset)
+                        struct anv_bo *bo, uint64_t offset)
 {
    const struct isl_device *isl_dev = &cmd_buffer->device->isl_dev;
